requirements.txt

flask
python-dotenv
tqdm
